Renamed test files from camelCase to train-case
