+++ test/services/callback-to-promise.test.js
