+++ test/services/pluck-query.test.js
