+++ test/services/remove-query.test.js
