+++ test/services/set-slug.test.js
